Rename keyboardShortcuts.js
