+++ src/main/features/core/keyboardShortcuts.js
